Feature: add source install instructions and cleanup readme
I added the instructions to use pip to install this package from source. I feel it is easy enough to explain does not make the readme any longer, so it is a nice touch. I also noticed that the readme had lines well over 100+ characters, making it difficult to edit on my monitor, so I added breaks in areas I felt helped organize the file.