change max line length to 120, fix indent on answers.py
This change adds poetry for dependency management as well as CI and pre-commit hooks. I needed to fix one issue in answers.py. Other than that, this change should not affect much existing code. It is important that everyone follows the new contributing instructions to guarantee new code added complies with our styling